use all PRB hooks at the top level in component, reorganise utility functions, improve naming
